refactor: message types
