+++ projects/ion/src/lib/message/message.component.ts
@@ -1,21 +1,7 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { IconType } from '../core/types/icon';
 
-export interface IonMessageProps {
-  label: string;
-  type?: MessageStatusType;
-  iconType?: IconType;
-}
-
-export type MessageStatusType =
-  | 'positive'
-  | 'negative_alert'
-  | 'negative_erro'
-  | 'warning'
-  | 'info'
-  | 'custom';
-
-export const icontypes = {
+const icontypes = {
   positive: 'check-solid',
   negative_alert: 'exclamation-solid',
   negative_erro: 'close-solid',
